requirements.txt
asgiref==3.5.0
black==22.3.0
certifi==2021.10.8
click==8.1.3
Django==4.0.4
django-filter==21.1
djangorestframework==3.13.1
flake8==4.0.1
isort==5.10.1
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
pytz==2022.1
sentry-sdk==1.5.12
sqlparse==0.4.2
tomli==2.0.1
urllib3==1.26.9
